+++ CHANGELOG.md
@@ -2,6 +2,30 @@
 
 All notable changes to this project will be documented in this file. See [commit-and-tag-version](https://github.com/absolute-version/commit-and-tag-version) for commit guidelines.
 
+## 0.2.0 (2026-06-18)
+
+
+### Dependencies
+
+* **deps:** Update DevKit to release-3.9.0
+
+
+### Features
+
+* add AVS support for issuing DNS role names
+* add different key extended use cases
+* add operator role in provisioned role
+
+
+### Bug Fixes
+
+* Preserve specific TCA error codes in AVS server
+
+
+### Documentation
+
+* Longer description for the purpose of AVS
+
 ## 0.1.0 (2026-06-12)